Security + een paar style changes

+++ database/api/activiteit.js
@@ -1,8 +1,7 @@
-import Database from "../database.js";
-const database = await Database.init();
+import { verifyAdmin } from "../../util/jwt-auth.js";
 const url = "/api/activiteit";
 
-export default function ActiviteitAPI(app) {
+export default function ActiviteitAPI(app, database) {
     app.get(url, async (req, res) => {
         if (req.query.id) {
             const activiteit = (await database.query("SELECT * FROM activiteit WHERE id = ? AND hidden = 0", [req.query.id]))[0];
@@ -14,6 +13,7 @@ export default function ActiviteitAPI(app) {
     });
 
     app.post(url, async (req, res) => {
+        if (!(await verifyAdmin(req.query.token))) return res.status(401).send("Unauthorized");
         if (req.body.naam && req.body.locatie && req.body.eten && req.body.omschrijving && req.body.begin && req.body.eind && req.body.kost && req.body.max && req.body.min && req.body.afbeelding && req.body.hidden !== undefined) {
             const activiteit = await database.query("INSERT INTO activiteit (naam, locatie, eten, omschrijving, begin, eind, kost, max, min, afbeelding, hidden) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", [req.body.naam, req.body.locatie, req.body.eten, req.body.omschrijving, req.body.begin, req.body.eind, req.body.kost, req.body.max, req.body.min, req.body.afbeelding, req.body.hidden ? 1 : 0]);
             res.send(activiteit);
@@ -23,6 +23,7 @@ export default function ActiviteitAPI(app) {
     });
 
     app.put(url, async (req, res) => {
+        if (!(await verifyAdmin(req.query.token))) return res.status(401).send("Unauthorized");
         if (req.body.id && req.body.naam && req.body.locatie && req.body.eten && req.body.omschrijving && req.body.begin && req.body.eind && req.body.kost && req.body.max && req.body.min && req.body.afbeelding && req.body.hidden !== undefined) {
             const activiteit = await database.query("UPDATE activiteit SET naam = ?, locatie = ?, eten = ?, omschrijving = ?, begin = ?, eind = ?, kost = ?, max = ?, min = ?, afbeelding = ?, hidden = ? WHERE id = ?", [req.body.naam, req.body.locatie, req.body.eten, req.body.omschrijving, req.body.begin, req.body.eind, req.body.kost, req.body.max, req.body.min, req.body.afbeelding, req.body.hidden ? 1 : 0, req.body.id]);
             res.send(activiteit);
@@ -32,6 +33,7 @@ export default function ActiviteitAPI(app) {
     });
 
     app.delete(url, async (req, res) => {
+        if (!(await verifyAdmin(req.query.token))) return res.status(401).send("Unauthorized");
         if (req.query.id) {
             const activiteit = await database.query("DELETE FROM activiteit WHERE id = ?", [req.query.id]);
             res.send(activiteit);
